Prevent uncaught exception on HttpListener.Close()
When the Close() method is called on an HttpListener instance, the
framework invokes the BeginGetContext method one last time and
previously because of how Nano was handling disposing resources some of
the variables used were disposed of. Nano will now properly return early
if the underlying resources are already disposed of. This really didn't
affect anything except some unit tests though.