v0.8 Release - Various fixes revealed thru testing
1. Introduced concept of an optional UnhandledRequestHandler to invoke
when a request was not handled by any other request or event handler.
This is of particular note when self-hosting as otherwise requests not
satisfied by a matched route / request handler would not be captureable.
With this addition the HttpListenerNanoServer will not inject a default
UnhandledRequestHandler implementation which returns an HTTP 404 NOT
FOUND for unhandled routes. The user may supply their own implementation
for logging purposes or to return their own custom response.
2. Introduced a few breaking changes: renamed a mispelled method
parameter, renamed a few methods, removed unused parameter.
3. Created a new implementation for the generic and non-generic Bind()
NanoContext extension methods for converting request parameters to a
particular type.
4. Modified the MethodRequestHandler parameter binding logic to not use
the lightweight type converter if the value looks like JSON. This was
causing JSON to be passed through as strings when bound to a type of
'object'.